@@ -1,25 +1,22 @@
-apply plugin: 'com.android.application'
-apply plugin: 'com.google.gms.google-services'  // Google Services plugin
+plugins {
+    id 'com.android.application'
+    id 'com.google.gms.google-services'
+}
 
 android {
-
-    compileOptions {
-        targetCompatibility = "8"
-        sourceCompatibility = "8"
-    }
-
-    compileSdkVersion 29
-    buildToolsVersion "29.0.3"
+    namespace 'com.ensias.healthcareapp'
+    compileSdk 35
 
     defaultConfig {
         applicationId "com.ensias.healthcareapp"
-        minSdkVersion 16
-        targetSdkVersion 29
+        minSdk 23
+        targetSdk 35
         versionCode 1
         versionName "1.0"
+
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         multiDexEnabled true
         vectorDrawables.useSupportLibrary = true
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
 
     buildTypes {
@@ -29,40 +26,61 @@ android {
         }
     }
 
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_17
+        targetCompatibility JavaVersion.VERSION_17
+    }
+
+    buildFeatures {
+        buildConfig true
+        viewBinding true
+    }
+}
+
+repositories {
+    google()
+    mavenCentral()
+    maven { url 'https://jitpack.io' }
 }
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
 
-    implementation 'androidx.appcompat:appcompat:1.1.0'
-    implementation "androidx.legacy:legacy-support-v4:1.0.0"
-    implementation "androidx.recyclerview:recyclerview:1.0.0"
-    implementation "androidx.viewpager:viewpager:1.0.0"
+    // Исправленные зависимости
+    implementation 'com.github.shuhart:StepView:1.5.1'
+    implementation 'com.github.Mulham-Raee:Horizontal-Calendar:1.3.2'
+    implementation 'io.github.pilgr:paperdb:2.7.2'
+
+
+    // AndroidX библиотеки
+    implementation 'androidx.appcompat:appcompat:1.7.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.recyclerview:recyclerview:1.4.0'
+    implementation 'androidx.viewpager:viewpager:1.1.0'
     implementation 'androidx.cardview:cardview:1.0.0'
-    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
-    testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.1'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
-    implementation 'com.google.firebase:firebase-auth:19.3.0'
-    implementation 'com.google.firebase:firebase-database:19.2.1'
-    implementation 'com.google.android.gms:play-services-auth:18.0.0'
-    implementation 'com.google.android.material:material:1.1.0'
-    implementation 'com.google.firebase:firebase-firestore:21.4.2'
-    implementation 'com.firebaseui:firebase-ui-storage:6.2.0'
-    implementation 'androidx.multidex:multidex:2.0.0'
-    implementation 'androidx.appcompat:appcompat:1.0.0'
-    implementation 'com.shuhart.stepview:stepview:1.5.1'
+    implementation 'androidx.constraintlayout:constraintlayout:2.2.1'
+    implementation 'androidx.multidex:multidex:2.0.1'
+
+    // Google Play Services
+    implementation 'com.google.android.material:material:1.12.0'
+    implementation 'com.google.android.gms:play-services-auth:21.3.0'
+
+    // Firebase
+    implementation 'com.google.firebase:firebase-auth:23.2.0'
+    implementation 'com.google.firebase:firebase-database:21.0.0'
+    implementation 'com.google.firebase:firebase-firestore:25.1.4'
+    implementation 'com.google.firebase:firebase-storage:21.0.1'
+    implementation 'com.firebaseui:firebase-ui-firestore:8.0.2'
+    implementation 'com.firebaseui:firebase-ui-storage:8.0.2'
+
+    // Дополнительные библиотеки
     implementation 'com.jaredrummler:material-spinner:1.3.1'
-    implementation 'devs.mulham.horizontalcalendar:horizontalcalendar:1.3.4'
-    implementation 'com.jakewharton:butterknife:10.0.0'
-    annotationProcessor 'com.jakewharton:butterknife-compiler:10.0.0'
     implementation 'com.github.d-max:spots-dialog:1.1@aar'
-    implementation 'io.paperdb:paperdb:2.7.1'
-    implementation 'com.google.code.gson:gson:2.8.5'
+    implementation 'com.google.code.gson:gson:2.10.1'
+    implementation 'com.squareup.picasso:picasso:2.71828'
 
-    implementation 'com.squareup.picasso:picasso:2.5.2'
-
-    implementation 'com.firebaseui:firebase-ui-firestore:4.1.0'
-    implementation 'com.google.firebase:firebase-storage:19.1.1'
-
+    // Тестовые библиотеки
+    testImplementation 'junit:junit:4.13.2'
+    androidTestImplementation 'androidx.test.ext:junit:1.2.1'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.6.1'
 }
