@@ -7,109 +7,88 @@ import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
-import android.widget.Button;
 import android.widget.DatePicker;
 import android.widget.Toast;
 
 import com.ensias.healthcareapp.Common.Common;
-import com.google.android.gms.common.SignInButton;
+import com.ensias.healthcareapp.databinding.ActivityDoctorHomeBinding;
 import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.firestore.FirebaseFirestore;
 
 import java.util.Calendar;
 
-import butterknife.BindView;
-import butterknife.ButterKnife;
-import butterknife.OnClick;
-import butterknife.Unbinder;
+public class DoctorHomeActivity extends AppCompatActivity implements DatePickerDialog.OnDateSetListener {
 
-public class DoctorHomeActivity extends AppCompatActivity implements DatePickerDialog.OnDateSetListener{
+    private ActivityDoctorHomeBinding binding; // ✅ правильный биндинг
+
     static String doc;
-    Button SignOutBtn2;
-    Button BtnRequst;
-    Button listPatients;
-    Button appointementBtn;
-    @OnClick(R.id.profile)
-    void profileBtnClick(){
-        Intent k = new Intent(DoctorHomeActivity.this, ProfileDoctorActivity.class);
-        startActivity(k);
-    }
-    Unbinder unbinder;
-
-    @OnClick(R.id.myCalendarBtn)
-    void myCalendarOnclick() {
-        Toast.makeText(this,"clicked",Toast.LENGTH_SHORT).show();
-        Intent k = new Intent(DoctorHomeActivity.this, MyCalendarDoctorActivity.class);
-        startActivity(k);
-    }
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_doctor_home); //ici layout de page d'acceuil MEDECIN
-        unbinder = ButterKnife.bind(this,this);
-        Common.CurreentDoctor = FirebaseAuth.getInstance().getCurrentUser().getEmail().toString();
+        binding = ActivityDoctorHomeBinding.inflate(getLayoutInflater()); // ✅ биндим layout активности
+        setContentView(binding.getRoot());
+
+        // Установить глобальные переменные
+        Common.CurreentDoctor = FirebaseAuth.getInstance().getCurrentUser().getEmail();
         Common.CurrentUserType = "doctor";
-        listPatients = findViewById(R.id.listPatients);
-        BtnRequst=findViewById(R.id.btnRequst);
-        SignOutBtn2=findViewById(R.id.signOutBtn);
-        appointementBtn = findViewById(R.id.appointement);
-        SignOutBtn2.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                FirebaseAuth.getInstance().signOut();
-                Intent intent = new Intent(getApplicationContext(), MainActivity.class);
-                intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
-                startActivity(intent);
-            }
-        });
-        BtnRequst.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent k = new Intent(DoctorHomeActivity.this, ConfirmedAppointmensActivity.class);
-                startActivity(k);
-            }
-        });
-        listPatients.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent k = new Intent(DoctorHomeActivity.this, MyPatientsActivity.class);
-                startActivity(k);
-            }
-        });
-        appointementBtn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-              // doc = FirebaseAuth.getInstance().getCurrentUser().getEmail().toString();
-                //showDatePickerDialog(v.getContext());
-                Intent k = new Intent(DoctorHomeActivity.this, DoctorAppointementActivity.class);
-                startActivity(k);
-            }
+
+        // Профиль
+        binding.profile.setOnClickListener(v -> {
+            startActivity(new Intent(this, ProfileDoctorActivity.class));
         });
 
+        // Календарь
+        binding.myCalendarBtn.setOnClickListener(v -> {
+            Toast.makeText(this, "clicked", Toast.LENGTH_SHORT).show();
+            startActivity(new Intent(this, MyCalendarDoctorActivity.class));
+        });
+
+        // Выход из аккаунта
+        binding.signOutBtn.setOnClickListener(v -> {
+            FirebaseAuth.getInstance().signOut();
+            Intent intent = new Intent(this, MainActivity.class);
+            intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
+            startActivity(intent);
+        });
+
+        // Подтвержденные записи
+        binding.btnRequst.setOnClickListener(v -> {
+            startActivity(new Intent(this, ConfirmedAppointmensActivity.class));
+        });
+
+        // Список пациентов
+        binding.listPatients.setOnClickListener(v -> {
+            startActivity(new Intent(this, MyPatientsActivity.class));
+        });
+
+        // Переход к приему
+        binding.appointement.setOnClickListener(v -> {
+            startActivity(new Intent(this, DoctorAppointementActivity.class));
+        });
     }
 
-    public void showDatePickerDialog(Context wf){
+    public void showDatePickerDialog(Context context) {
         DatePickerDialog datePickerDialog = new DatePickerDialog(
-                wf,
+                context,
                 this,
                 Calendar.getInstance().get(Calendar.YEAR),
                 Calendar.getInstance().get(Calendar.MONTH),
-                Calendar.getInstance().get(Calendar.DAY_OF_MONTH));
+                Calendar.getInstance().get(Calendar.DAY_OF_MONTH)
+        );
         datePickerDialog.show();
     }
 
     @Override
     public void onDateSet(DatePicker view, int year, int month, int dayOfMonth) {
         String date = "month_day_year: " + month + "_" + dayOfMonth + "_" + year;
-        openPage(view.getContext(),doc,date);
+        openPage(view.getContext(), doc, date);
     }
 
-    private void openPage(Context wf, String d,String day){
-        Intent i = new Intent(wf, AppointementActivity.class);
-        i.putExtra("key1",d+"");
-        i.putExtra("key2",day);
-        i.putExtra("key3","doctor");
-        wf.startActivity(i);
+    private void openPage(Context context, String doctorEmail, String day) {
+        Intent intent = new Intent(context, AppointementActivity.class);
+        intent.putExtra("key1", doctorEmail);
+        intent.putExtra("key2", day);
+        intent.putExtra("key3", "doctor");
+        context.startActivity(intent);
     }
 }
