@@ -94,7 +94,7 @@ public class EditProfilePatientActivity extends AppCompatActivity {
         pathReference.getDownloadUrl().addOnSuccessListener(new OnSuccessListener<Uri>() {
             @Override
             public void onSuccess(Uri uri) {
-                Picasso.with(EditProfilePatientActivity.this)
+                Picasso.get()
                         .load(uri)
                         .placeholder(R.drawable.doctor)
                         .fit()
@@ -172,7 +172,7 @@ public class EditProfilePatientActivity extends AppCompatActivity {
         if (requestCode == PICK_IMAGE_REQUEST && resultCode == RESULT_OK
                 && data != null && data.getData() != null) {
             uriImage = data.getData();
-            Picasso.with(this).load(uriImage).into(profileImage);
+            Picasso.get().load(uriImage).into(profileImage);
         }
     }
 
