@@ -3,164 +3,131 @@ package com.ensias.healthcareapp;
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.recyclerview.widget.GridLayoutManager;
-import androidx.recyclerview.widget.RecyclerView;
 
 import android.app.AlertDialog;
-import android.os.Binder;
 import android.os.Bundle;
 import android.widget.Toast;
 
 import com.ensias.healthcareapp.Common.Common;
 import com.ensias.healthcareapp.Interface.ITimeSlotLoadListener;
 import com.ensias.healthcareapp.adapter.MyTimeSlotAdapter;
+import com.ensias.healthcareapp.databinding.ActivityMyCalendarDoctorBinding;
 import com.ensias.healthcareapp.model.TimeSlot;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.Task;
-import com.google.firebase.firestore.CollectionReference;
-import com.google.firebase.firestore.DocumentReference;
-import com.google.firebase.firestore.DocumentSnapshot;
-import com.google.firebase.firestore.FirebaseFirestore;
-import com.google.firebase.firestore.QueryDocumentSnapshot;
-import com.google.firebase.firestore.QuerySnapshot;
+import com.google.firebase.firestore.*;
 
 import java.text.SimpleDateFormat;
-import java.util.ArrayList;
-import java.util.Calendar;
-import java.util.List;
-import java.util.logging.Formatter;
-import java.util.logging.SimpleFormatter;
+import java.util.*;
 
-import butterknife.BindView;
-import butterknife.ButterKnife;
 import devs.mulham.horizontalcalendar.HorizontalCalendar;
-import devs.mulham.horizontalcalendar.HorizontalCalendarView;
 import devs.mulham.horizontalcalendar.utils.HorizontalCalendarListener;
 import dmax.dialog.SpotsDialog;
 
 public class MyCalendarDoctorActivity extends AppCompatActivity implements ITimeSlotLoadListener {
 
-    DocumentReference doctorDoc;
+    private ActivityMyCalendarDoctorBinding binding;
 
-    ITimeSlotLoadListener iTimeSlotLoadListener;
-    android.app.AlertDialog alertDialog;
-    @BindView(R.id.recycle_time_slot2)
-    RecyclerView recycler_time_slot;
-    @BindView(R.id.calendarView2)
-    HorizontalCalendarView calendarView;
-    private SimpleDateFormat simpleDateFormat = new SimpleDateFormat("dd_MM_yyyy");
+    private DocumentReference doctorDoc;
+    private ITimeSlotLoadListener iTimeSlotLoadListener;
+    private AlertDialog alertDialog;
+    private final SimpleDateFormat simpleDateFormat = new SimpleDateFormat("dd_MM_yyyy");
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_my_calendar_doctor);
-        ButterKnife.bind(this);
-        init();
+        binding = ActivityMyCalendarDoctorBinding.inflate(getLayoutInflater());
+        setContentView(binding.getRoot());
 
+        init();
     }
 
     private void init() {
         iTimeSlotLoadListener = this;
-        alertDialog = new SpotsDialog.Builder().setCancelable(false).setContext(this)
-                .build();
+        alertDialog = new SpotsDialog.Builder().setCancelable(false).setContext(this).build();
+
         Calendar date = Calendar.getInstance();
-        date.add(Calendar.DATE,0);
-        loadAvailabelTimeSlotOfDoctor(Common.CurreentDoctor,simpleDateFormat.format(date.getTime()));
-        recycler_time_slot.setHasFixedSize(true);
-        GridLayoutManager gridLayoutManager = new GridLayoutManager(this,3);
-        recycler_time_slot.setLayoutManager(gridLayoutManager);
+        loadAvailabelTimeSlotOfDoctor(Common.CurreentDoctor, simpleDateFormat.format(date.getTime()));
+
+        binding.recycleTimeSlot2.setHasFixedSize(true);
+        binding.recycleTimeSlot2.setLayoutManager(new GridLayoutManager(this, 3));
 
         Calendar startDate = Calendar.getInstance();
-        startDate.add(Calendar.DATE,0);
         Calendar endDate = Calendar.getInstance();
-        endDate.add(Calendar.DATE,5);
-        HorizontalCalendar horizontalCalendar = new HorizontalCalendar.Builder(this,R.id.calendarView2)
-                .range(startDate,endDate)
+        endDate.add(Calendar.DATE, 5);
+
+        HorizontalCalendar calendar = new HorizontalCalendar.Builder(this, binding.calendarView2.getId())
+                .range(startDate, endDate)
                 .datesNumberOnScreen(1)
-                .mode(HorizontalCalendar.Mode.DAYS)
                 .defaultSelectedDate(startDate)
                 .build();
-        horizontalCalendar.setCalendarListener(new HorizontalCalendarListener() {
+
+        calendar.setCalendarListener(new HorizontalCalendarListener() {
             @Override
             public void onDateSelected(Calendar date, int position) {
-                if(Common.currentDate.getTimeInMillis() != date.getTimeInMillis()){
+                if (Common.currentDate.getTimeInMillis() != date.getTimeInMillis()) {
                     Common.currentDate = date;
-                    loadAvailabelTimeSlotOfDoctor(Common.CurreentDoctor,simpleDateFormat.format(date.getTime()));
-
+                    loadAvailabelTimeSlotOfDoctor(Common.CurreentDoctor, simpleDateFormat.format(date.getTime()));
                 }
-
             }
         });
-
     }
 
     private void loadAvailabelTimeSlotOfDoctor(String curreentDoctor, String bookDate) {
-    alertDialog.show();
+        alertDialog.show();
 
         doctorDoc = FirebaseFirestore.getInstance()
                 .collection("Doctor")
-                .document(Common.CurreentDoctor);
-        doctorDoc.get().addOnCompleteListener(new OnCompleteListener<DocumentSnapshot>() {
-            @Override
-            public void onComplete(@NonNull Task<DocumentSnapshot> task) {
-                if(task.isSuccessful())
-                {
-                    DocumentSnapshot documentSnapshot = task.getResult();
-                    if(documentSnapshot.exists()){
-                        CollectionReference date =FirebaseFirestore.getInstance()
-                                .collection("Doctor")
-                                .document(Common.CurreentDoctor)
-                                .collection(bookDate);
+                .document(curreentDoctor);
 
-                        date.get().addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
-                            @Override
-                            public void onComplete(@NonNull Task<QuerySnapshot> task) {
-                                if (task.isSuccessful())
-                                {
-                                    QuerySnapshot querySnapshot = task.getResult();
-                                    if (querySnapshot.isEmpty())
-                                    {
-                                        iTimeSlotLoadListener.onTimeSlotLoadEmpty();
-                                    }else {
-                                        List<TimeSlot> timeSlots = new ArrayList<>();
-                                        for (QueryDocumentSnapshot document:task.getResult())
-                                            timeSlots.add(document.toObject(TimeSlot.class));
-                                        iTimeSlotLoadListener.onTimeSlotLoadSuccess(timeSlots);
+        doctorDoc.get().addOnCompleteListener(task -> {
+            if (task.isSuccessful()) {
+                DocumentSnapshot docSnap = task.getResult();
+                if (docSnap != null && docSnap.exists()) {
+                    CollectionReference dateRef = FirebaseFirestore.getInstance()
+                            .collection("Doctor")
+                            .document(curreentDoctor)
+                            .collection(bookDate);
+
+                    dateRef.get()
+                            .addOnCompleteListener(snapshotTask -> {
+                                if (snapshotTask.isSuccessful()) {
+                                    List<TimeSlot> timeSlots = new ArrayList<>();
+                                    for (QueryDocumentSnapshot doc : snapshotTask.getResult()) {
+                                        timeSlots.add(doc.toObject(TimeSlot.class));
                                     }
 
+                                    if (timeSlots.isEmpty()) {
+                                        iTimeSlotLoadListener.onTimeSlotLoadEmpty();
+                                    } else {
+                                        iTimeSlotLoadListener.onTimeSlotLoadSuccess(timeSlots);
+                                    }
                                 }
-                            }
-                        }).addOnFailureListener(new OnFailureListener() {
-                            @Override
-                            public void onFailure(@NonNull Exception e) {
-                                iTimeSlotLoadListener.onTimeSlotLoadFailed(e.getMessage());
-                            }
-                        });
-                    }
+                            })
+                            .addOnFailureListener(e -> iTimeSlotLoadListener.onTimeSlotLoadFailed(e.getMessage()));
                 }
             }
         });
-
     }
-
 
     @Override
     public void onTimeSlotLoadSuccess(List<TimeSlot> timeSlotList) {
-        MyTimeSlotAdapter adapter = new MyTimeSlotAdapter(this,timeSlotList);
-        recycler_time_slot.setAdapter(adapter);
+        MyTimeSlotAdapter adapter = new MyTimeSlotAdapter(this, timeSlotList);
+        binding.recycleTimeSlot2.setAdapter(adapter);
         alertDialog.dismiss();
     }
 
     @Override
     public void onTimeSlotLoadFailed(String message) {
-        Toast.makeText(this,message,Toast.LENGTH_SHORT).show();
+        Toast.makeText(this, message, Toast.LENGTH_SHORT).show();
         alertDialog.dismiss();
     }
 
     @Override
     public void onTimeSlotLoadEmpty() {
         MyTimeSlotAdapter adapter = new MyTimeSlotAdapter(this);
-        recycler_time_slot.setAdapter(adapter);
+        binding.recycleTimeSlot2.setAdapter(adapter);
         alertDialog.dismiss();
     }
 }
