@@ -2,80 +2,63 @@ package com.ensias.healthcareapp.adapter;
 
 import android.net.Uri;
 import android.view.LayoutInflater;
-import android.view.View;
 import android.view.ViewGroup;
-import android.widget.Button;
-import android.widget.ImageView;
-import android.widget.TextView;
-
-import com.ensias.healthcareapp.R;
-import com.ensias.healthcareapp.model.ApointementInformation;
-import com.ensias.healthcareapp.model.Doctor;
-import com.firebase.ui.firestore.FirestoreRecyclerAdapter;
-import com.firebase.ui.firestore.FirestoreRecyclerOptions;
-import com.google.android.gms.tasks.OnFailureListener;
-import com.google.android.gms.tasks.OnSuccessListener;
-import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.storage.FirebaseStorage;
-import com.google.firebase.storage.StorageReference;
-import com.squareup.picasso.Picasso;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
+import com.ensias.healthcareapp.databinding.ConfirmedAppointementsItemBinding;
+import com.ensias.healthcareapp.model.ApointementInformation;
+import com.firebase.ui.firestore.FirestoreRecyclerAdapter;
+import com.firebase.ui.firestore.FirestoreRecyclerOptions;
+import com.google.android.gms.tasks.OnFailureListener;
+import com.google.android.gms.tasks.OnSuccessListener;
+import com.google.firebase.storage.FirebaseStorage;
+import com.google.firebase.storage.StorageReference;
+import com.squareup.picasso.Picasso;
+
 public class ConfirmedAppointmentsAdapter extends FirestoreRecyclerAdapter<ApointementInformation, ConfirmedAppointmentsAdapter.ConfirmedAppointmentsHolder> {
-    StorageReference pathReference ;
+
     public ConfirmedAppointmentsAdapter(@NonNull FirestoreRecyclerOptions<ApointementInformation> options) {
         super(options);
     }
 
     @Override
-    protected void onBindViewHolder(@NonNull ConfirmedAppointmentsHolder confirmedAppointmentsHolder, int position, @NonNull final ApointementInformation apointementInformation) {
-        confirmedAppointmentsHolder.dateAppointement.setText(apointementInformation.getTime());
-        confirmedAppointmentsHolder.patientName.setText(apointementInformation.getPatientName());
-        confirmedAppointmentsHolder.appointementType.setText(apointementInformation.getApointementType());
+    protected void onBindViewHolder(@NonNull ConfirmedAppointmentsHolder holder, int position, @NonNull ApointementInformation appointment) {
+        holder.binding.appointementDate.setText(appointment.getTime());
+        holder.binding.patientName.setText(appointment.getPatientName());
+        holder.binding.appointementType.setText(appointment.getApointementType());
 
-        String imageId = apointementInformation.getPatientId()+".jpg"; //add a title image
-        pathReference = FirebaseStorage.getInstance().getReference().child("DoctorProfile/"+ imageId); //storage the image
-        pathReference.getDownloadUrl().addOnSuccessListener(new OnSuccessListener<Uri>() {
-            @Override
-            public void onSuccess(Uri uri) {
-                Picasso.with(confirmedAppointmentsHolder.patientImage.getContext())
+        String imageId = appointment.getPatientId() + ".jpg";
+        StorageReference pathReference = FirebaseStorage.getInstance().getReference().child("DoctorProfile/" + imageId);
+
+        pathReference.getDownloadUrl()
+                .addOnSuccessListener(uri -> Picasso.get()
                         .load(uri)
-                        .placeholder(R.mipmap.ic_launcher)
+                        .placeholder(android.R.drawable.ic_menu_report_image)
                         .fit()
                         .centerCrop()
-                        .into(confirmedAppointmentsHolder.patientImage);//Image location
-
-                // profileImage.setImageURI(uri);
-            }
-        }).addOnFailureListener(new OnFailureListener() {
-            @Override
-            public void onFailure(@NonNull Exception exception) {
-                // Handle any errors
-            }
-        });
-
+                        .into(holder.binding.patientImage))
+                .addOnFailureListener(exception -> {
+                    // Ошибка загрузки изображения (можно добавить лог или плейсхолдер)
+                });
     }
 
     @NonNull
     @Override
-    public ConfirmedAppointmentsAdapter.ConfirmedAppointmentsHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
-        View v = LayoutInflater.from(parent.getContext()).inflate(R.layout.confirmed_appointements_item, parent, false);
-        return new ConfirmedAppointmentsAdapter.ConfirmedAppointmentsHolder(v);
+    public ConfirmedAppointmentsHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
+        ConfirmedAppointementsItemBinding binding = ConfirmedAppointementsItemBinding.inflate(
+                LayoutInflater.from(parent.getContext()), parent, false
+        );
+        return new ConfirmedAppointmentsHolder(binding);
     }
 
-    class ConfirmedAppointmentsHolder extends RecyclerView.ViewHolder{
-        TextView dateAppointement;
-        TextView patientName;
-        TextView appointementType;
-        ImageView patientImage;
-        public ConfirmedAppointmentsHolder(@NonNull View itemView) {
-            super(itemView);
-            dateAppointement = itemView.findViewById(R.id.appointement_date);
-            patientName = itemView.findViewById(R.id.patient_name);
-            appointementType = itemView.findViewById(R.id.appointement_type);
-            patientImage = itemView.findViewById(R.id.patient_image);
+    static class ConfirmedAppointmentsHolder extends RecyclerView.ViewHolder {
+        ConfirmedAppointementsItemBinding binding;
+
+        public ConfirmedAppointmentsHolder(@NonNull ConfirmedAppointementsItemBinding binding) {
+            super(binding.getRoot());
+            this.binding = binding;
         }
     }
 }
