@@ -20,29 +20,14 @@ import androidx.recyclerview.widget.RecyclerView;
 
 import com.ensias.healthcareapp.Common.Common;
 import com.ensias.healthcareapp.Interface.ITimeSlotLoadListener;
-import com.ensias.healthcareapp.R;
 import com.ensias.healthcareapp.adapter.MyTimeSlotAdapter;
+import com.ensias.healthcareapp.databinding.FragmentBookingStepTwoBinding;
 import com.ensias.healthcareapp.model.TimeSlot;
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.OnFailureListener;
-import com.google.android.gms.tasks.Task;
-import com.google.firebase.firestore.CollectionReference;
-import com.google.firebase.firestore.DocumentReference;
-import com.google.firebase.firestore.DocumentSnapshot;
-import com.google.firebase.firestore.FirebaseFirestore;
-import com.google.firebase.firestore.QueryDocumentSnapshot;
-import com.google.firebase.firestore.QuerySnapshot;
-
-import org.xml.sax.helpers.NamespaceSupport;
+import com.google.firebase.firestore.*;
 
 import java.text.SimpleDateFormat;
-import java.util.ArrayList;
-import java.util.Calendar;
-import java.util.List;
+import java.util.*;
 
-import butterknife.BindView;
-import butterknife.ButterKnife;
-import butterknife.Unbinder;
 import devs.mulham.horizontalcalendar.HorizontalCalendar;
 import devs.mulham.horizontalcalendar.HorizontalCalendarView;
 import devs.mulham.horizontalcalendar.utils.HorizontalCalendarListener;
@@ -50,96 +35,65 @@ import dmax.dialog.SpotsDialog;
 
 public class BookingStep2Fragment extends Fragment implements ITimeSlotLoadListener {
 
-    DocumentReference doctorDoc;
-    ITimeSlotLoadListener iTimeSlotLoadListener;
-    AlertDialog dialog;
+    private FragmentBookingStepTwoBinding binding;
+    private ITimeSlotLoadListener iTimeSlotLoadListener;
+    private AlertDialog dialog;
+    private LocalBroadcastManager localBroadcastManager;
+    private SimpleDateFormat simpleDateFormat;
 
-    Unbinder unbinder;
-    LocalBroadcastManager localBroadcastManager;
-
-
-    @BindView(R.id.recycle_time_slot)
-    RecyclerView recycler_time_slot;
-    @BindView(R.id.calendarView)
-    HorizontalCalendarView calendarView;
-    SimpleDateFormat simpleDateFormat;
-
-    BroadcastReceiver displayTimeSlot = new BroadcastReceiver() {
+    private final BroadcastReceiver displayTimeSlot = new BroadcastReceiver() {
         @Override
         public void onReceive(Context context, Intent intent) {
             Calendar date = Calendar.getInstance();
-            date.add(Calendar.DATE,0);
-            loadAvailabelTimeSlotOfDoctor(Common.CurreentDoctor,simpleDateFormat.format(date.getTime()));
-
+            date.add(Calendar.DATE, 0);
+            loadAvailableTimeSlotOfDoctor(Common.CurreentDoctor, simpleDateFormat.format(date.getTime()));
         }
     };
 
-    private void loadAvailabelTimeSlotOfDoctor(String doctorId, final String bookDate) {
+    private void loadAvailableTimeSlotOfDoctor(String doctorId, final String bookDate) {
         dialog.show();
-
-        doctorDoc = FirebaseFirestore.getInstance()
+        FirebaseFirestore.getInstance()
                 .collection("Doctor")
-                .document(Common.CurreentDoctor);
-        doctorDoc.get().addOnCompleteListener(new OnCompleteListener<DocumentSnapshot>() {
-            @Override
-            public void onComplete(@NonNull Task<DocumentSnapshot> task) {
-                if(task.isSuccessful())
-                {
-                    DocumentSnapshot documentSnapshot = task.getResult();
-                    if(documentSnapshot.exists()){
-                        CollectionReference date =FirebaseFirestore.getInstance()
+                .document(doctorId)
+                .get()
+                .addOnSuccessListener(documentSnapshot -> {
+                    if (documentSnapshot.exists()) {
+                        FirebaseFirestore.getInstance()
                                 .collection("Doctor")
-                                .document(Common.CurreentDoctor)
-                                .collection(bookDate);
-
-                        date.get().addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
-                            @Override
-                            public void onComplete(@NonNull Task<QuerySnapshot> task) {
-                                if (task.isSuccessful())
-                                {
-                                    QuerySnapshot querySnapshot = task.getResult();
-                                    if (querySnapshot.isEmpty())
-                                    {
+                                .document(doctorId)
+                                .collection(bookDate)
+                                .get()
+                                .addOnSuccessListener(querySnapshot -> {
+                                    if (querySnapshot.isEmpty()) {
                                         iTimeSlotLoadListener.onTimeSlotLoadEmpty();
-                                    }else {
+                                    } else {
                                         List<TimeSlot> timeSlots = new ArrayList<>();
-                                        for (QueryDocumentSnapshot document:task.getResult())
-                                            timeSlots.add(document.toObject(TimeSlot.class));
+                                        for (QueryDocumentSnapshot doc : querySnapshot)
+                                            timeSlots.add(doc.toObject(TimeSlot.class));
                                         iTimeSlotLoadListener.onTimeSlotLoadSuccess(timeSlots);
                                     }
-
-                                }
-                            }
-                        }).addOnFailureListener(new OnFailureListener() {
-                            @Override
-                            public void onFailure(@NonNull Exception e) {
-                                iTimeSlotLoadListener.onTimeSlotLoadFailed(e.getMessage());
-                            }
-                        });
+                                })
+                                .addOnFailureListener(e -> iTimeSlotLoadListener.onTimeSlotLoadFailed(e.getMessage()));
                     }
-                }
-            }
-        });
+                });
     }
 
     static BookingStep2Fragment instance;
-    public  static  BookingStep2Fragment getInstance(){
-        if(instance == null )
+
+    public static BookingStep2Fragment getInstance() {
+        if (instance == null)
             instance = new BookingStep2Fragment();
         return instance;
     }
+
     @Override
     public void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-
         iTimeSlotLoadListener = this;
-
-        localBroadcastManager = LocalBroadcastManager.getInstance(getContext());
-        localBroadcastManager.registerReceiver(displayTimeSlot,new IntentFilter(Common.KEY_DISPLAY_TIME_SLOT));
-        simpleDateFormat = new SimpleDateFormat("dd_MM_yyyy");
-
+        localBroadcastManager = LocalBroadcastManager.getInstance(requireContext());
+        localBroadcastManager.registerReceiver(displayTimeSlot, new IntentFilter(Common.KEY_DISPLAY_TIME_SLOT));
+        simpleDateFormat = new SimpleDateFormat("dd_MM_yyyy", Locale.getDefault());
         dialog = new SpotsDialog.Builder().setContext(getContext()).setCancelable(false).build();
-
     }
 
     @Override
@@ -151,64 +105,52 @@ public class BookingStep2Fragment extends Fragment implements ITimeSlotLoadListe
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-         super.onCreateView(inflater, container, savedInstanceState);
-        View itemView = inflater.inflate(R.layout.fragment_booking_step_two,container,false);
-        unbinder = ButterKnife.bind(this,itemView);
-
-        init(itemView);
-        loadAvailabelTimeSlotOfDoctor(Common.CurreentDoctor,simpleDateFormat.format(Common.currentDate.getTime()));
-
-        return itemView;
-
+        binding = FragmentBookingStepTwoBinding.inflate(inflater, container, false);
+        init();
+        loadAvailableTimeSlotOfDoctor(Common.CurreentDoctor, simpleDateFormat.format(Common.currentDate.getTime()));
+        return binding.getRoot();
     }
 
-    private void init(View itemView) {
-        recycler_time_slot.setHasFixedSize(true);
-        GridLayoutManager gridLayoutManager = new GridLayoutManager(getActivity(),3);
-        recycler_time_slot.setLayoutManager(gridLayoutManager);
-        //recycler_time_slot.addItemDecoration(new SpaceI);
+    private void init() {
+        binding.recycleTimeSlot.setHasFixedSize(true);
+        binding.recycleTimeSlot.setLayoutManager(new GridLayoutManager(getContext(), 3));
 
         Calendar startDate = Calendar.getInstance();
-        startDate.add(Calendar.DATE,0);
         Calendar endDate = Calendar.getInstance();
-        endDate.add(Calendar.DATE,5);
-        HorizontalCalendar horizontalCalendar = new HorizontalCalendar.Builder(itemView,R.id.calendarView)
-                .range(startDate,endDate)
+        endDate.add(Calendar.DATE, 5);
+
+        HorizontalCalendar horizontalCalendar = new HorizontalCalendar.Builder(binding.getRoot(), binding.calendarView.getId())
+                .range(startDate, endDate)
                 .datesNumberOnScreen(1)
-                .mode(HorizontalCalendar.Mode.DAYS)
                 .defaultSelectedDate(startDate)
                 .build();
 
         horizontalCalendar.setCalendarListener(new HorizontalCalendarListener() {
             @Override
             public void onDateSelected(Calendar date, int position) {
-                if(Common.currentDate.getTimeInMillis() != date.getTimeInMillis()){
+                if (Common.currentDate.getTimeInMillis() != date.getTimeInMillis()) {
                     Common.currentDate = date;
-                    loadAvailabelTimeSlotOfDoctor(Common.CurreentDoctor,simpleDateFormat.format(date.getTime()));
-
+                    loadAvailableTimeSlotOfDoctor(Common.CurreentDoctor, simpleDateFormat.format(date.getTime()));
                 }
-
             }
         });
     }
 
     @Override
     public void onTimeSlotLoadSuccess(List<TimeSlot> timeSlotList) {
-        MyTimeSlotAdapter adapter = new MyTimeSlotAdapter(getContext(),timeSlotList);
-        recycler_time_slot.setAdapter(adapter);
+        binding.recycleTimeSlot.setAdapter(new MyTimeSlotAdapter(getContext(), timeSlotList));
         dialog.dismiss();
     }
 
     @Override
     public void onTimeSlotLoadFailed(String message) {
-        Toast.makeText(getContext(),message,Toast.LENGTH_SHORT).show();
+        Toast.makeText(getContext(), message, Toast.LENGTH_SHORT).show();
         dialog.dismiss();
     }
 
     @Override
     public void onTimeSlotLoadEmpty() {
-        MyTimeSlotAdapter adapter = new MyTimeSlotAdapter(getContext());
-        recycler_time_slot.setAdapter(adapter);
+        binding.recycleTimeSlot.setAdapter(new MyTimeSlotAdapter(getContext()));
         dialog.dismiss();
     }
 }
